workshop/taxi-pipeline/pyproject.toml
[project]
name = "taxi-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "ibis-framework[duckdb]>=12.0.0",
    "marimo>=0.20.2",
    "plotly>=6.5.2",
]
